Improve user interface of dropdown menu
The dropdown menu was displaying the links inline because the display property on the links was only on when the menu items were hovered. This edit makes the display property of the links constantly equal to "block" thus never allowing the text to go inline and creating a more seamless transition.